separated physical from mental agent representation
git-svn-id: file:///srv/local/data/cleanRepo/trunk@18248 a0517364-e873-4167-b1c8-416daa5a6b91

+++ playgrounds/gregor/src/main/java/playground/gregor/sim2d_v2/random/XORShiftRandom.java
@@ -17,6 +17,6 @@ public class XORShiftRandom {
 
 	public double nextDouble() {
 		long l = randomLong();
-		return (double)l/Long.MAX_VALUE;
+		return (double)l/ (double)Long.MAX_VALUE;
 	}
 }
